requirements.txt
mlflow==2.10.2
dvc[gs]==3.32.0
langchain==0.1.7
pinecone-client==2.2.4
unstructured==0.11.2
gpt4all==2.0.2
tiktoken==0.5.2
evaluate==0.4.1
textstat==0.7.3
rouge-score==0.1.2
bentoml==1.1.10
bentoctl==0.4.0
python-dotenv~=1.0.0
openllm==0.4.35
streamlit==1.31.1
uvicorn==0.24.0.post1
fastapi==0.109.2
langserve==0.0.41
